refactor: Application을 부트스트랩 전용으로 분리
